Fixed copyright header and code format
