+++ src/main/java/com/rapid7/client/dcerpc/mssamr/messages/SamrOpenDomainResponse.java
@@ -1,8 +1,21 @@
-/***************************************************************************
- * COPYRIGHT (C) 2017, Rapid7 LLC, Boston, MA, USA.
- * All rights reserved. This material contains unpublished, copyrighted
- * work including confidential and proprietary information of Rapid7.
- **************************************************************************/
+/**
+ * Copyright 2017, Rapid7, Inc.
+ *
+ * License: BSD-3-clause
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * * Redistributions of source code must retain the above copyright notice,
+ *   this list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright
+ *   notice, this list of conditions and the following disclaimer in the
+ *   documentation and/or other materials provided with the distribution.
+ *
+ * * Neither the name of the copyright holder nor the names of its contributors
+ *   may be used to endorse or promote products derived from this software
+ *   without specific prior written permission.
+ */
 package com.rapid7.client.dcerpc.mssamr.messages;
 
 import com.rapid7.client.dcerpc.messages.HandleResponse;
@@ -12,6 +25,6 @@ public class SamrOpenDomainResponse extends HandleResponse<DomainHandle> {
 
     @Override
     protected DomainHandle initHandle() {
-        return new DomainHandle();
+	return new DomainHandle();
     }
 }
